files/requirements.txt
boto3
docopt
PyYAML